Document build script use and manual testing
Explains how the build script is used, for setting up the initial application, and
that changes to node_modules and SCSS files will need to be rebuilt and the output
commited.

That isn't ideal, and those files should be treated as generated, built during
setup and/or deploy, rather than commited to the repo itself. Fixing that is
outside the scope of this change, so i'm going to great an issue on the repo
to track that suggestion.

This also adds instructions on manually testin the done/ pages locally, without
having to go through the payment provider. This was extrmely useful while
fixing the application CSS, and I suspect will be useful to someone else in
the future.