Cleanup broken CSS references that were 404ing
These seem to be a holdover from before govuk_template was being used
in a programantic fashion. They're using a different naming convetion
and none of the files referenced would exist.

Also the main application CSS (in assetPath not govukTemplateAssetPath)
was not being served to IE8 or older, which is bad.

In particular the print file was visibly 404ing, as it wasn't obscured
by the IE conditional include.

All of these, save the main application CSS sans-conditional, can be
removed